Improve compatibilty with OSX, wine preapacs.exe
Add parameter PREPACS_BIN in section INGEST of accode.ini to specify how to run
the binary prepacs.exe
Note that line 313 and 315 unreadable caracters have been deleted those lines
were commented, need to make them compatible with OSX

Minor modification in the rest of the code to integrate this modification.

Modification tested on OSX but not on Windows.